# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Mealswapp API
  version: 0.3.0
  description: Shared gateway source of truth for Phase 03 authentication and gateway contracts.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/v1/admin/classifications:
    get:
      operationId: getApiV1AdminClassifications
      summary: List one active global classification hierarchy
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ClassificationKindQuery"
      responses:
        "200":
          description: Active classifications for one global kind.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClassificationCollectionEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-009 TagManager classification creation, update, and guarded deletion.
  /api/v1/admin/classifications/{classification}:
    post:
      operationId: postApiV1AdminClassification
      summary: Create one global classification
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ClassificationKindPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminClassificationRequest"
      responses:
        "201":
          $ref: "#/components/responses/AdminClassification"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    put:
      operationId: putApiV1AdminClassification
      summary: Rename or reparent one global classification
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ClassificationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminClassificationRequest"
      responses:
        "200":
          $ref: "#/components/responses/AdminClassification"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteApiV1AdminClassification
      summary: Soft-delete one unused global classification
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/ClassificationId"
      responses:
        "204":
          description: Classification deleted with no client-visible audit snapshot.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-009 UserAdminPanel privacy-minimized bounded lookup.
...
components:
  parameters:
...
    ClassificationKindQuery:
      name: kind
      in: query
      required: true
      schema:
        type: string
        enum: [food_category, culinary_role]
    ClassificationKindPath:
      name: classification
      in: path
      required: true
      schema:
        type: string
        enum: [food_category, culinary_role]
    ClassificationId:
      name: classification
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mealswapp_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: mealswapp_refresh
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
...
    Error:
      description: User-safe gateway error envelope with no client-visible audit snapshot, raw provider payload, or internal diagnostics.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    # Implements DESIGN-004 JobStatusTracker and DESIGN-010 RateLimiter shared admission contract.
    TooManyRequests:
      description: Retryable rate or admission rejection with a positive whole-second delay.
      headers:
        Retry-After:
          required: true
          description: Positive whole seconds before the caller should retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AdminClassification:
      description: One global classification hierarchy node without audit state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminClassificationEnvelope"
...
  schemas:
    AppError:
      type: object
      additionalProperties: false
      required: [category, code, message, retryable]
      properties:
        category:
          type: string
          enum: [validation, auth, entitlement, security, network, timeout, server, dependency, rate_limit, unknown]
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 1
          maxLength: 120
    # Implements DESIGN-009 AdminController audit-safe error boundary.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, error]
      properties:
        status:
          type: string
          const: error
        requestId:
          type: string
          maxLength: 120
        error:
          $ref: "#/components/schemas/AppError"
...
    AdminClassificationRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Runtime trims and collapses whitespace, normalizes to NFC, and accepts only Unicode letters, marks, digits, spaces, and punctuation; maximum length is 120 Unicode code points after normalization.
        parentId:
          type: [string, "null"]
          format: uuid
    AdminClassification:
      type: object
      additionalProperties: false
      required: [id, name, kind]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Runtime trims and collapses whitespace, normalizes to NFC, and accepts only Unicode letters, marks, digits, spaces, and punctuation; maximum length is 120 Unicode code points after normalization.
        kind:
          type: string
          enum: [food_category, culinary_role]
        parentId:
          type: string
          format: uuid
    AdminClassificationEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, data]
      properties:
        status:
          type: string
          const: ok
        requestId:
          type: string
        data:
          type: object
          additionalProperties: false
          required: [classification]
          properties:
            classification:
              $ref: "#/components/schemas/AdminClassification"
    AdminClassificationCollectionEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, data]
      properties:
        status:
          type: string
          const: ok
        requestId:
          type: string
        data:
          type: object
          additionalProperties: false
          required: [classifications]
          properties:
            classifications:
              type: array
              maxItems: 1000
              items:
                $ref: "#/components/schemas/AdminClassification"
    # Implements DESIGN-009 UserAdminPanel bounded privacy-minimized response.
...